move FixedQueue to a special package
